+++ example/text_indexing_example.dart
@@ -7,34 +7,41 @@
 import 'package:text_indexing/text_indexing.dart';
 
 /// Two examples using the indexers in this package are provided:
-/// - [_inMemoryIndexerExample] is a simple example of a [InMemoryIndexer]
+/// - [_inMemoryIndexerExample] is a simple example of a [TextIndexer.inMemory]
 ///   indexing the [textData] dataset; and
-/// - [_persistedIndexerExample] is a simple example of a [AsyncIndexer]
+/// - [_persistedIndexerExample] is a simple example of a [TextIndexer.async]
 ///   indexing the [textData] dataset.
 void main() async {
   //
 
+  const searchPhrase = 'stock market tesla EV battery';
+
   // Run a simple example of the [InMemoryIndexer] on the [textData] dataset.
-  await _inMemoryIndexerExample(textData);
+  await _inMemoryIndexerExample(textData, searchPhrase);
 
   //  Run a simple example of the [AsyncIndexer] on the [textData] dataset.
-  await _persistedIndexerExample(jsonData);
+  await _persistedIndexerExample(jsonData, searchPhrase);
 
   //
 }
 
-/// A simple example of the [InMemoryIndexer] on the [documents] dataset:
+/// A simple example of the [TextIndexer.inMemory] on the [documents] dataset:
 /// - initialize the [Dictionary];
 /// - initialize the [Postings];
-/// - initialize a [InMemoryIndexer];
-/// - listen to the [InMemoryIndexer.postingsStream], printing the
+/// - initialize a [TextIndexer];
+/// - listen to the [TextIndexer.postingsStream], printing the
 ///   emitted postings for each indexed document;
-/// - iterate through the sample data;
-/// - index each document, adding/updating terms in the [Dictionary]
-///   and postings in the [Postings] ; and
+/// - iterate through the sample data, indexing each document in turn; and
 /// - print the top 5 most popular [Dictionary.terms].
-Future<void> _inMemoryIndexerExample(Map<String, String> documents) async {
+Future<void> _inMemoryIndexerExample(
+    Map<String, String> documents, String searchPhrase) async {
   //
+
+// initialize a Set for the vocabulary state
+  final termsSet = <String>{};
+
+  // initialize a Set for the document ids state
+  final docsSet = <String>{};
 
   // - initialize the [Dictionary]
   final dictionary = <String, int>{};
@@ -46,11 +53,18 @@ Future<void> _inMemoryIndexerExample(Map<String, String> documents) async {
   final indexer =
       TextIndexer.inMemory(dictionary: dictionary, postings: postings);
 
+  /// - tokenize a phrase into searh terms
+  final searchTerms =
+      (await indexer.index.analyzer.tokenize(searchPhrase)).tokens.terms;
+
+  // listen to the indexer.postingsStream and print the postings count for each term.
   // listen to the indexer.postingsStream and print the postings count for each term.
   indexer.postingsStream.listen((event) {
-    for (final term in event.entries) {
-      print('Term ${term.key} has ${term.value.length} postings');
+    for (final termPosting in event.entries) {
+      termsSet.add(termPosting.key);
+      docsSet.add(termPosting.value.entries.first.key);
     }
+    print('Indexed ${termsSet.length} terms from ${docsSet.length} documents.');
   });
 
   // - iterate through the sample data
@@ -62,93 +76,71 @@ Future<void> _inMemoryIndexerExample(Map<String, String> documents) async {
   // wait for stream elements to complete printing
   await Future.delayed(const Duration(milliseconds: 250));
 
-  // print the 5 most popuplar terms with their frequencies
-  var terms = dictionary.toList(TermSortStrategy.byFrequency);
-  if (terms.length > 5) {
-    terms = terms.sublist(0, 5);
-  }
-  for (final term in terms) {
-    print('${term.term}: ${term.dFt}');
-  }
+  // print the document term frequencies for each term in searchTerms
+  await _printTermFrequencyPostings(indexer.index, searchTerms);
+
+  // print the statistics for each term in [searchTerms].
+  await _printTermStats(indexer.index, searchTerms);
 }
 
-/// A simple test of the [AsyncIndexer] on a small JSON dataset using a
+/// A simple test of the [TextIndexer.async] on a small JSON dataset using a
 /// simulated persisted index repository with 50 millisecond latency on
 /// read/write operations to [Dictionary] and [Postings] hashmaps:
-/// - initialize the [_TestIndex()];
-/// - initialize a [AsyncIndexer];
-/// - listen to the [AsyncIndexer.postingsStream], printing the
-///   emitted postings for each indexed document;
-/// - iterate through the JSON documents;
-/// - index each document, adding/updating terms in the [_TestIndex.dictionary]
-///   and postings in the [_TestIndex.postings] ; and
-/// - print the top 5 most popular [_TestIndex.dictionary.terms].
-Future<void> _persistedIndexerExample(Map<String, JSON> documents) async {
+/// - initialize the [_TestIndex];
+/// - initialize a [TextIndexer];
+/// - listen to the [TextIndexer.postingsStream];
+/// - iterate through the JSON documents and index each document; and
+/// - print the statistics on 5 search terms.
+Future<void> _persistedIndexerExample(
+    Map<String, JSON> documents, String searchPhrase) async {
   //
+
+// initialize a Set for the vocabulary state
+  final termsSet = <String>{};
+
+  // initialize a Set for the document ids state
+  final docsSet = <String>{};
 
   // - initialize a [_TestIndex()]
   final index = _TestIndex();
 
   /// - tokenize a phrase into searh terms
   final searchTerms =
-      (await index.analyzer.tokenize('stock market tesla EV battery'))
-          .tokens
-          .terms;
+      (await index.analyzer.tokenize(searchPhrase)).tokens.terms;
 
   // - initialize a [AsyncIndexer]
   final indexer = TextIndexer.index(index: index);
 
   // listen to the indexer.postingsStream and print the postings count for each term.
   indexer.postingsStream.listen((event) {
-    if (event.isNotEmpty) {
-      final docPostings = event.values.first;
-      if (docPostings.isNotEmpty) {
-        print(
-            'Postings received for document  {${docPostings.entries.first.key}}');
-      }
+    for (final termPosting in event.entries) {
+      termsSet.add(termPosting.key);
+      docsSet.add(termPosting.value.entries.first.key);
     }
+    print('Indexed ${termsSet.length} terms from ${docsSet.length} documents.');
   });
 
-  // - define the FieldNames to be indexed
-  final fields = ['name', 'description', 'hashTags', 'publicationDate'];
+  print('Indexed ${termsSet.length} terms from ${docsSet.length} documents.');
 
   // - iterate through the sample data
   await indexer.indexCollection(jsonData);
 
   // wait for stream elements to complete printing
   await Future.delayed(const Duration(milliseconds: 250));
-// get the 5 most popular terms with their frequencies
-  var terms = index.dictionary.toList(TermSortStrategy.byFrequency);
-  if (terms.length > 5) {
-    terms = terms.sublist(0, 5);
-  }
 
-  // print the 5 most popular terms with their frequencies
-  print('_______________________________________________________');
-  print('DOCUMENT FREQUENCY (top 5 terms)');
-  for (final term in terms) {
-    print('Term: "${term.term}":   dFt: ${term.dFt}');
-  }
+  // print the document term frequencies for each term in searchTerms
+  await _printTermFrequencyPostings(index, searchTerms);
 
-  /// get the inverse term frequency index for the search terms
-  final iDftIndex = await index.getIdFtIndex(searchTerms);
+  // print the statistics for each term in [searchTerms].
+  await _printTermStats(index, searchTerms);
+}
 
-  // get the inverse document frequency index.
-  var idftIndex = iDftIndex.toList(TermSortStrategy.byFrequency);
-  if (idftIndex.length > 5) {
-    idftIndex = idftIndex.sublist(0, 5);
-  }
-  // print the 5 terms with the highest inverse document frequency.
-  print('_______________________________________________________');
-  print('INVERSE DOCUMENT FREQUENCY (search terms)');
-  for (final term in idftIndex) {
-    print('Term: "${term.term}":    iDFt ${term.iDFt.toStringAsFixed(2)}');
-  }
-
+/// Print the document term frequencies for each term in [searchTerms].
+Future<void> _printTermFrequencyPostings(
+    InvertedIndex index, Iterable<String> searchTerms) async {
   /// get the document term frequency postings for the search terms
   final tfPostings = await index.getFtdPostings(searchTerms, 2);
 
-  /// print the term frequency postings for the search terms
   print('_______________________________________________________');
   print('TERM DOCUMENT FREQUENCY (search terms, minimum dFt = 2)');
   for (final e in tfPostings.entries) {
@@ -156,6 +148,42 @@ Future<void> _persistedIndexerExample(Map<String, JSON> documents) async {
     for (final posting in e.value.entries) {
       print('   DocId: {${posting.key}}:    dFt: [${posting.value}]');
     }
+  }
+}
+
+/// Print the statistics for each term in [searchTerms].
+Future<void> _printTermStats(
+    InvertedIndex index, Iterable<String> searchTerms) async {
+  //
+
+  // get the inverse term frequency index for the searchTerms
+  final iDftIndex = await index.getIdFtIndex(searchTerms);
+
+  // get the term frequency in the corpus of the searchTerms
+  final tFtIndex = await index.getTfIndex(searchTerms);
+
+  // get the dictionary for searchTerms
+  final dictionary = await index.getDictionary(searchTerms);
+
+  // print the headings
+  print(''.padLeft(80, '_'));
+  print('DICTIONARY STATISTICS (search terms, minimum dFt = 2)');
+  print(''.padLeft(80, '-'));
+  print('${'Term'.padRight(10)}'
+      '${'Term Frequency'.toString().padLeft(20)}'
+      '${'Document Frequency'.toString().padLeft(20)}'
+      '${'Inverse Document Frequency'.toString().padLeft(30)}');
+  print(''.padLeft(80, '-'));
+
+  // print the statistics
+  for (final term in searchTerms) {
+    final df = dictionary[term] ?? 0;
+    final idf = iDftIndex[term] ?? 0.0;
+    final tf = tFtIndex[term] ?? 0;
+    print('${term.padRight(10)}'
+        '${tf.toString().padLeft(20)}'
+        '${df.toString().padLeft(20)}'
+        '${idf.toStringAsFixed(2).padLeft(4, '0').padLeft(30)}');
   }
 }